Close Steam before Shortcut Gen
All of this was done pretty quickly so there may be some edge cases I am not accounting for.  process_names contains a list of keywords that I found to appear in the names of steam programs. I am not sure which ones need to be closed, so that probably requires editing.